attempt to limit the creation of monit scripts only for application deployments that exist

+++ grantoo_resque_monit/recipes/default.rb
@@ -40,6 +40,11 @@ end
 sleep(5.0)
 
 node[:deploy].each do |application, deploy|
+  if !File.exists?(deploy[:deploy_to])
+    Chef::Log.info "skipping deploy_to #{deploy[:deploy_to]} -- does not exist"
+    next
+  end
+
   Dir.glob(File.join(node[:monit][:conf_dir], "grantoo_resque*.monitrc")).each do |f|
     File.delete(f)
   end
